Remove store argument from get_artifact_repository() (#1138)
Remove the store parameter from ArtifactRepositoryRegistry.get_artifact_repository() by having DBFSArtifactRepository internally resolve the host credentials via tracking.utils._get_store().

This leads to a circular import: tracking.utils (for artifact/model methods) imports get_artifact_repository imports DBFSArtifactRepository imports tracking.utils._get_store. Python 2 did not like this (Python 3.5+ supports circular imports better), so moved out the artifact-related methods from tracking.utils to tracking.artifact_utils. A lot of the changes in the PR are simple refactors due to this move.

This change will allow Artifact Repositories to be used for downloading artifacts outside the Tracking context, e.g. in <model-type>.load_model.
