+++ mlflow/sklearn.py
@@ -242,7 +242,7 @@ def load_model(path, run_id=None):
     >>> predictions = sk_model.predict(pandas_df)
     """
     if run_id is not None:
-        path = mlflow.tracking.utils._get_model_log_dir(model_name=path, run_id=run_id)
+        path = mlflow.tracking.artifact_utils._get_model_log_dir(model_name=path, run_id=run_id)
     path = os.path.abspath(path)
     flavor_conf = _get_flavor_configuration(model_path=path, flavor_name=FLAVOR_NAME)
     sklearn_model_artifacts_path = os.path.join(path, flavor_conf['pickled_model'])
