+++ mlflow/store/artifact_repository_registry.py
@@ -11,7 +11,6 @@ from mlflow.store.sftp_artifact_repo import SFTPArtifactRepository
 from mlflow.store.dbfs_artifact_repo import DbfsArtifactRepository
 from mlflow.store.s3_artifact_repo import S3ArtifactRepository
 from mlflow.store.local_artifact_repo import LocalArtifactRepository
-from mlflow.store.rest_store import RestStore
 
 
 class ArtifactRepositoryRegistry:
@@ -47,33 +46,26 @@ class ArtifactRepositoryRegistry:
                     stacklevel=2
                 )
 
-    def get_artifact_repository(self, artifact_uri, store=None):
+    def get_artifact_repository(self, artifact_uri):
         """Get an artifact repository from the registry based on the scheme of artifact_uri
 
         :param store_uri: The store URI. This URI is used to select which artifact repository
                           implementation to instantiate and is passed to the
                           constructor of the implementation.
-        :param store: Instance of `AbstractStore`. This is currently only used to get the host
-                      credentials when instantiating a DBFS-backed artifact repository.
 
         :return: An instance of `mlflow.store.ArtifactRepository` that fulfills the artifact URI
                  requirements.
         """
         scheme = urllib.parse.urlparse(artifact_uri).scheme
         repository = self._registry.get(scheme)
-        if scheme == "dbfs" and repository is not None:
-            if not isinstance(store, RestStore):
-                raise MlflowException('`store` must be an instance of RestStore.')
-            return repository(artifact_uri, store.get_host_creds)
-        elif repository is not None:
-            return repository(artifact_uri)
-        else:
+        if repository is None:
             raise MlflowException(
                 "Could not find a registered artifact repository for: {}. "
                 "Currently registered schemes are: {}".format(
                     artifact_uri, list(self._registry.keys())
                 )
             )
+        return repository(artifact_uri)
 
 
 _artifact_repository_registry = ArtifactRepositoryRegistry()
@@ -90,16 +82,14 @@ _artifact_repository_registry.register('hdfs', HdfsArtifactRepository)
 _artifact_repository_registry.register_entrypoints()
 
 
-def get_artifact_repository(artifact_uri, store=None):
+def get_artifact_repository(artifact_uri):
     """Get an artifact repository from the registry based on the scheme of artifact_uri
 
     :param store_uri: The store URI. This URI is used to select which artifact repository
                       implementation to instantiate and is passed to the
                       constructor of the implementation.
-    :param store: Instance of `AbstractStore`. This is currently only used to get the host
-                  credentials when instantiating a DBFS-backed artifact repository.
 
     :return: An instance of `mlflow.store.ArtifactRepository` that fulfills the artifact URI
              requirements.
     """
-    return _artifact_repository_registry.get_artifact_repository(artifact_uri, store)
+    return _artifact_repository_registry.get_artifact_repository(artifact_uri)
