+++ mlflow/tracking/utils.py
@@ -8,9 +8,7 @@ import entrypoints
 from six.moves import urllib
 
 from mlflow.exceptions import MlflowException
-from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
 from mlflow.store import DEFAULT_LOCAL_FILE_AND_ARTIFACT_PATH
-from mlflow.store.artifact_repository_registry import get_artifact_repository
 from mlflow.store.dbmodels.db_types import DATABASE_ENGINES
 from mlflow.store.file_store import FileStore
 from mlflow.store.rest_store import RestStore
@@ -72,58 +70,6 @@ def get_tracking_uri():
         return env.get_env(_TRACKING_URI_ENV_VAR)
     else:
         return os.path.abspath(DEFAULT_LOCAL_FILE_AND_ARTIFACT_PATH)
-
-
-def get_artifact_uri(run_id, artifact_path=None):
-    """
-    Get the absolute URI of the specified artifact in the specified run. If `path` is not specified,
-    the artifact root URI of the specified run will be returned; calls to ``log_artifact``
-    and ``log_artifacts`` write artifact(s) to subdirectories of the artifact root URI.
-
-    :param run_id: The ID of the run for which to obtain an absolute artifact URI.
-    :param artifact_path: The run-relative artifact path. For example,
-                          ``path/to/artifact``. If unspecified, the artifact root URI for the
-                          specified run will be returned.
-    :return: An *absolute* URI referring to the specified artifact or the specified run's artifact
-             root. For example, if an artifact path is provided and the specified run uses an
-             S3-backed  store, this may be a uri of the form
-             ``s3://<bucket_name>/path/to/artifact/root/path/to/artifact``. If an artifact path
-             is not provided and the specified run uses an S3-backed store, this may be a URI of
-             the form ``s3://<bucket_name>/path/to/artifact/root``.
-    """
-    if not run_id:
-        raise MlflowException(
-                message="A run_id must be specified in order to obtain an artifact uri!",
-                error_code=INVALID_PARAMETER_VALUE)
-
-    store = _get_store()
-    run = store.get_run(run_id)
-    if artifact_path is None:
-        return run.info.artifact_uri
-    else:
-        # Path separators may not be consistent across all artifact repositories. Therefore, when
-        # joining the run's artifact root directory with the artifact's relative path, we use the
-        # path module defined by the appropriate artifact repository
-        artifact_path_module =\
-            get_artifact_repository(run.info.artifact_uri, store).get_path_module()
-        return artifact_path_module.join(run.info.artifact_uri, artifact_path)
-
-
-def _download_artifact_from_uri(artifact_uri, output_path=None):
-    """
-    :param artifact_uri: The *absolute* URI of the artifact to download.
-    :param output_path: The local filesystem path to which to download the artifact. If unspecified,
-                        a local output path will be created.
-    """
-    store = _get_store(artifact_uri=artifact_uri)
-    artifact_path_module =\
-        get_artifact_repository(artifact_uri, store).get_path_module()
-    artifact_src_dir = artifact_path_module.dirname(artifact_uri)
-    artifact_src_relative_path = artifact_path_module.basename(artifact_uri)
-    artifact_repo = get_artifact_repository(
-            artifact_uri=artifact_src_dir, store=store)
-    return artifact_repo.download_artifacts(
-            artifact_path=artifact_src_relative_path, dst_path=output_path)
 
 
 def _is_local_uri(uri):
@@ -272,15 +218,7 @@ def _get_store(store_uri=None, artifact_uri=None):
     return _tracking_store_registry.get_store(store_uri, artifact_uri)
 
 
-def _get_model_log_dir(model_name, run_id):
-    if not run_id:
-        raise Exception("Must specify a run_id to get logging directory for a model.")
-    store = _get_store()
-    run = store.get_run(run_id)
-    artifact_repo = get_artifact_repository(run.info.artifact_uri, store)
-    return artifact_repo.download_artifacts(model_name)
-
-
+# TODO(sueann): move to a projects utils module
 def _get_git_url_if_present(uri):
     """
     Return the path git_uri#sub_directory if the URI passed is a local path that's part of
