+++ tests/sagemaker/test_deployment.py
@@ -19,7 +19,7 @@ from mlflow.exceptions import MlflowException
 from mlflow.models import Model
 from mlflow.protos.databricks_pb2 import ErrorCode, RESOURCE_DOES_NOT_EXIST, INVALID_PARAMETER_VALUE
 from mlflow.protos.databricks_pb2 import INTERNAL_ERROR
-from mlflow.tracking.utils import _get_model_log_dir
+from mlflow.tracking.artifact_utils import _get_model_log_dir
 
 from tests.sagemaker.mock import mock_sagemaker, Endpoint, EndpointOperation
